requirements.txt
Flask-SQLAlchemy===2.4.1 
Flask-Migrate==2.5.2 
Flask-Script==2.0.6
flask-restful==0.3.8
flask-jsonpify==1.5.0